Compatibility with commons-vfs2:2.4.1
I actually change one of the methods in the sand box original connector, to use a deprecated method for extracting the url schema.
This allows us to be backward compatible with commons-vfs2:2.4.1

For smb the deprecated method works fine, since the standard manager contains both schemas : "smb1" and "smb", it will always parse the schema correctly.